+++ cmd/task/unreleased.go
@@ -3,12 +3,12 @@ package task
 import (
 	"fmt"
 
-	"release/cmd/commandkey"
-	"release/cmd/flagkey"
-	"release/cmd/task/flag"
-	"release/files"
-	"release/parser"
-	"release/utils"
+	"github.com/tomodian/release/cmd/commandkey"
+	"github.com/tomodian/release/cmd/flagkey"
+	"github.com/tomodian/release/cmd/task/flag"
+	"github.com/tomodian/release/files"
+	"github.com/tomodian/release/parser"
+	"github.com/tomodian/release/utils"
 
 	"github.com/urfave/cli/v2"
 )